+++ ejo-core.php
@@ -3,7 +3,7 @@
  * Plugin Name:         EJO Core
  * Plugin URI:          http://github.com/ejoweb/ejo-core
  * Description:         EJOweb core functionalities for theme development. Including some nifty debug tools.
- * Version:             0.5.2
+ * Version:             0.5.3
  * Author:              Erik Joling
  * Author URI:          http://www.ejoweb.nl/
  * GitHub Plugin URI:   https://github.com/EJOweb/ejo-core
@@ -26,7 +26,7 @@
 final class EJO_Core 
 {
     //* Version number of this plugin
-    public static $version = '0.5.2';
+    public static $version = '0.5.3';
 
     //* Holds the instance of this class.
     protected static $_instance = null;
@@ -63,6 +63,10 @@ final class EJO_Core
         //* Add shortcodes
         add_action( 'plugins_loaded', array( $this, 'add_shortcodes' ) );
 
+        //* Load Cleaners
+        add_action( 'plugins_loaded', array( $this, 'add_cleaners' ) );
+        // add_action( 'after_setup_theme', array( $this, 'add_cleaners' ), 99 );
+
         //* Add EJOcore Option page to Wordpress Option menu
         add_action( 'admin_menu', array( $this, 'register_options_page' ) );
     }
@@ -96,6 +100,16 @@ final class EJO_Core
         include_once( self::$dir . 'includes/dev-functions/image-size-crop.php' );
     }
 
+    //* Add Cleaner Functions
+    public function add_cleaners() 
+    {
+        //* Disable XML-RPC
+        include_once( self::$dir . 'includes/cleaners/disable-xmlrpc.php' );
+
+        //* Clean <head> links
+        include_once( self::$dir . 'includes/cleaners/head-links.php' );
+
+    }    
   
     //* Add Theme Support Tools
     public function add_theme_tools()